Created a backend (nodejs + typescript) and also splited into frontend and backend respectfuly
