+++ backend/.gitignore
@@ -1,0 +1,3 @@
+.env
+package-lock.json
+/node_modules
