+++ backend/package.json
@@ -1,0 +1,19 @@
+{
+	"name": "backend",
+	"version": "1.0.0",
+	"description": "",
+	"main": "index.js",
+	"scripts": {
+		"start": "nodemon index.ts",
+		"test": "echo \"Error: no test specified\" && exit 1"
+	},
+	"keywords": [],
+	"author": "",
+	"license": "ISC",
+	"devDependencies": {
+		"@types/node": "^16.9.1",
+		"nodemon": "^2.0.12",
+		"ts-node": "^10.2.1",
+		"typescript": "^4.4.3"
+	}
+}
