+++ frontend/package.json
@@ -1,19 +1,15 @@
 {
-  "name": "open_source_club",
+  "name": "frontend",
   "version": "0.1.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
-    "format": "prettier --write \"./pages/**/*.{js,jsx,css}\"",
-    "check-format": "prettier --check \"./pages/**/*.{js,jsx,css}\""
+    "lint": "next lint"
   },
   "dependencies": {
-    "axios": "^0.21.3",
-    "next": "^11.1.2",
-    "prettier": "^2.3.2",
+    "next": "11.1.2",
     "react": "17.0.2",
     "react-dom": "17.0.2"
   },
