--- pages/duke.js
@@ -1,23 +1,0 @@
-import React from 'react';
-import { useState, useEffect } from 'react';
-import axios from 'axios';
-const TEST_URL = 'https://api.openweathermap.org/data/2.5/weather?q=LasVegas&appid=';
-
-export default function Sardorbek() {
-    const [temp, setTemp] = useState(null);
-
-    useEffect(() => {
-        axios.get(TEST_URL + process.env.NEXT_PUBLIC_OPEN_WEATHER_API_KEY).then((res) => {
-            setTemp(res.data.main.temp);
-            console.log(res.data);
-        });
-    }, []);
-
-    return (
-        <div>
-            <div>Duke</div>
-            <div>Las Vegas</div>
-            <div>Temp: {Math.ceil(temp - 273.15)}</div>
-        </div>
-    );
-}
